[BUGFIX] Fetch all fe_user fields in findEntitiesWithIpAuthentication
The fe_users fields fetched in findEntitiesWithIpAuthentication are used in the authentication process, e.g. to determine user groups. If only uid and pid is fetched, the authentication service doesn't consider the user as authenticated even if a user session is generated. So the user is only correctly authenticated in the next request, e.g. after a redirect.